[FIX] mrp: extra moves set correct quantity_done and product_uom_qty
Before, when we consumed 3 instead of 1 in a production order, it
gave a traceback.  It was because it created moves
with negative quantities.  So, something was wrong with
the quantity_done/product_uom_qty set when creating the extra moves

When creating an extra move because of the procurement,
but you still need to split afterwards (e.g. procurement 10, production order 12, you do 11),
then you should have a difference between quantity_done and product_uom_qty
in the extra move and not try to change it on the original.
That way, we make sure we pass the correct quantities.