[FIX] web_editor: don't paste two times
When pasting, the paste event is cancelled and done back with the
corrected data.

This works fine on firefox, but this can have unwanted results on other
browser like pasting two times or copying the end of the line at the
beginning.

This commit replace completely the summernote functionality with our
small implementation of pasting corrected text.

closes #14043
opw-690814