[FIX] website_quote: tokenized link in portal
When a user has access to a quote (because he is follower) but he is not
in the same company than the customer, an access error can occur on the
the rendering of the partner field.

We use the tokenized link instead, which will read the SO as a superuser
to avoid this situation.

opw-691740