[FIX] calendar: invitation template lang
- Create a calendar event
- Add invitees

The invitees receive a mail so they can accept/decline. When clicking,
the page is always displayed in the language of the administrator.

Since the `request` object is not directly associated to the attendee,
the template is not rendered in his language. Since (i) the template
should not be rendered lazily (see comment in code) and (ii)
`request.render()` does not allow the rendering in an arbitraty
language, we do a manual call to `render_template`.

opw-694718