[FIX] web: prevent crash in line chart with 1 record
Since commit 86252428846607cd5f4b18c, the graph view crashed when
displaying a line chart with only one record.  The reason is that the
method display_line can return undefined (if no line chart can be
displayed).  In that case, there is no reasy to try to display its
tooltip.

This commit simply protect against the case where it is undefined.